chore: move NetworkType to standalone file
